pyproject.toml
[project]
name = "udbnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "h5py>=3.12.1",
    "ipython>=8.31.0",
    "matplotlib>=3.10.0",
    "numpy<2",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]
